Split out Client code
